updated to jclouds snapshots so that we can remove integer workarounds for ttl

+++ providers/denominator-ultradns/src/main/java/denominator/ultradns/UltraDNSPredicates.java
@@ -28,7 +28,7 @@ final class UltraDNSPredicates {
 
         @Override
         public boolean apply(ResourceRecord in) {
-            return typeValue == in.getType().intValue();
+            return typeValue == in.getType();
         }
 
         @Override
